pyproject.toml
[project]
name = "specforge"
version = "2.0.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "scrapling>=0.2.0",
    "playwright>=1.49.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sse-starlette>=2.1.0",
    "pillow>=10.4.0",
    "jsonschema>=4.23.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    "aiohttp>=3.10.0",
    "tenacity>=9.0.0",
]

[project.scripts]
specforge = "specforge.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/specforge"]
